Updated load-bpf.c so that it actually loads eBPF bytecode.
This change introduces `bpf_trace_printk.py`, which I created as a
utility to leverage the Python BPF library to compile a C program
and dump the eBPF bytecode. I was then able to take this bytecode and
inline it into the `load-bpf.c` source file so I could create the
`bpf_insn[]` programmatically.

Although the compiled version of `load-bpf.c` runs, it does not really
do anything yet because it does not attach the kprobe.