Improve print_list_of_instructions() debug function.
The idea is that we can change the `bpf_text`
in `bpf_trace_printk.py` to whatever we like and then
use this to dump the bytecode.
This is basically the same as what `llvm-objdump -S` does,
though it gives us a little more control over the
output.

Here is the new output:

```
$ python bpf_trace_printk.py
0 180100006f6d2072 # r1 = 0x0a74737572206d6f
1 000000007573740a
2 7b1ae8ff00000000 # *(u64 *) (r10 + -24) = r1
3 1801000068656c6c # r1 = 0x7266206f6c6c6568
4 000000006f206672
5 7b1ae0ff00000000 # *(u64 *) (r10 + -32) = r1
6 b701000000000000 # r1 = 0
7 731af0ff00000000 # *(u8 *) (r10 + -16) = r1
8 bfa1000000000000 # r1 = r10
9 07010000e0ffffff # r1 += -32
10 b702000011000000 # r2 = 17
11 8500000006000000 # call bpf_trace_printk()
12 b700000000000000 # r0 = 0
13 9500000000000000 # exit
```

Admittedly, it might be better to simply use the bcc tools
to dump the ELF file and then just use `llvm-objdump`, but
I haven't found the magic incantation in libbpf or elsewhere
to do that yet.