Support --pid and --tid options.
Note this is a non-trivial change because it requires
generating the bytecode for three variants of `trace_entry()`:

* show all opens
* filter opens by PID
* filter opens by TID

For the filter variants, when generating the bytecode generation
function, we must insert placeholders for the PID and TID variables.
The logic to do this is a little hacky, but seems good enough.