Attempt to put the kprobe and kretprobe code in one BPF program.
I don't know what the `offset` argument can be used for in
`bpf_attach_kprobe()` in libbpf, so I tried concatenating all
of the bytecode in the kprobe and kretprobe programs into a
single program, loading it via `bpf_prog_load()`, and then
specifying the non-zero offset into the BPF instruction array
for the other program. (If this worked, this would save a bpf(2)
syscall in opensnoop!)

Unfortunately, this failed with:

```
bpf: Failed to load program: Invalid argument
unreachable insn 82
```

Basically, the BPF validator realizes that the second program
that was concatenated is unreachable from the first and
complains, so `bpf_prog_load()` fails and I cannot even attempt
to call `bpf_attach_kprobe()` with the offset. This jives
with this comment from `linux/kernel/bpf/verifier.c` (the third bullet):

```
* The first pass is depth-first-search to check that the program is a DAG.
* It rejects the following programs:
* - larger than BPF_MAXINSNS insns
* - if loop is present (detected via back-edge)
* - unreachable insns exist (shouldn't be a forest. program = one function)
* - out of bounds or malformed jumps
```